refactor: rename Config to Session
